Update to how the injector handles error messages when starting bots.
Remove jugu from blacklist

reenable last in base behavior for devastation

fix typo in error message matching string

Make the bot close the process when disconnected or lag locked instead of trying to log back in.  Logging back in will be handled by the injector for better control of timing and spam prevention.